Update introduction.md - Round #1
So I changed the story up a little bit to maybe add some fun animations to the beginning of the chapter. You'll see. :)

I wanted to call it "Gbay" but that already exists, apparently. lol

I also made some tiny tweaks to the wording, so double check me and make sure what I wrote is accurate.